Restructure mixins
