+++ src/main/java/dev/emi/trinkets/mixin/client/CreativeInventoryScreenMixin.java
@@ -1,52 +1,48 @@
-package dev.emi.trinkets.mixin;
+package dev.emi.trinkets.mixin.client;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import com.mojang.blaze3d.platform.GlStateManager;
 import com.mojang.blaze3d.systems.RenderSystem;
 
+import dev.emi.trinkets.mixin.SlotMixin;
 import net.minecraft.client.gui.screen.ingame.AbstractInventoryScreen;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.screen.PlayerScreenHandler;
+import net.minecraft.screen.ScreenHandler;
 import net.minecraft.screen.slot.Slot;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import dev.emi.trinkets.TrinketInventoryRenderer;
-import dev.emi.trinkets.TrinketSlot;
 import dev.emi.trinkets.TrinketsClient;
-import dev.emi.trinkets.api.SlotGroups;
+import dev.emi.trinkets.api.TrinketInventory;
 import dev.emi.trinkets.api.TrinketSlots;
 import dev.emi.trinkets.api.TrinketSlots.SlotGroup;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.minecraft.client.gui.screen.ingame.InventoryScreen;
-import net.minecraft.client.gui.screen.recipebook.RecipeBookProvider;
+import net.minecraft.client.gui.screen.ingame.CreativeInventoryScreen;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.item.ItemGroup;
 import net.minecraft.text.Text;
-import net.minecraft.util.Identifier;
 
 /**
- * Rendering, logic, etc. for slot groups and extra slots
+ * Trinkets slots in the creative inventory, noticably grosser than the survival inventory due to slot restrictions
  */
 @Environment(EnvType.CLIENT)
-@Mixin(InventoryScreen.class)
-public abstract class InventoryScreenMixin extends AbstractInventoryScreen<PlayerScreenHandler> implements RecipeBookProvider {
+@Mixin(CreativeInventoryScreen.class)
+public abstract class CreativeInventoryScreenMixin extends AbstractInventoryScreen<CreativeInventoryScreen.CreativeScreenHandler> {
 	@Shadow
-	private float mouseX;
-	@Shadow
-	private float mouseY;
-	@Shadow
-	private static final Identifier RECIPE_BUTTON_TEXTURE = new Identifier("textures/gui/recipe_button.png");
-	
-	private List<TrinketSlot> invSlots;
+	private static int selectedTab;
 
-	public InventoryScreenMixin(PlayerScreenHandler screenHandler, PlayerInventory inventory, Text text) {
+	private List<Slot> creativeSlots = new ArrayList<Slot>();
+	private int mouseX, mouseY;
+
+	public CreativeInventoryScreenMixin(CreativeInventoryScreen.CreativeScreenHandler screenHandler, PlayerInventory inventory, Text text) {
 		super(screenHandler, inventory, text);
 	}
 
@@ -56,44 +52,46 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 	}
 
 	@Inject(at = @At("TAIL"), method = "init")
-	public void init(CallbackInfo info){
+	public void init(CallbackInfo info) {
+		if(creativeSlots.size() == 0) return;
 		TrinketsClient.displayEquipped = 0;
-		invSlots = new ArrayList<>();
-		for (Slot slot: this.getScreenHandler().slots) {
-			if (slot instanceof TrinketSlot) {
-				TrinketSlot ts = (TrinketSlot) slot;
-				invSlots.add(ts);
-				if (!ts.keepVisible) {
-					((SlotMixin) (Object) slot).setXPosition(Integer.MIN_VALUE);
-				} else {
-					((SlotMixin) (Object) ts).setXPosition(getGroupX(TrinketSlots.getSlotFromName(ts.group, ts.slot).getSlotGroup()) + 1);
-					((SlotMixin) (Object) ts).setYPosition(getGroupY(TrinketSlots.getSlotFromName(ts.group, ts.slot).getSlotGroup()) + 1);
-				}
+		List<TrinketSlots.Slot> trinketSlots = TrinketSlots.getAllSlots();
+		for (int i = 0; i < trinketSlots.size(); i++) {
+			if (!trinketSlots.get(i).getSlotGroup().onReal && trinketSlots.get(i).getSlotGroup().slots.get(0).equals(trinketSlots.get(i))) {
+				((SlotMixin) (Object) creativeSlots.get(i)).setXPosition(getGroupX(trinketSlots.get(i).getSlotGroup()) + 1);
+				((SlotMixin) (Object) creativeSlots.get(i)).setYPosition(getGroupY(trinketSlots.get(i).getSlotGroup()) + 1);
+			} else {
+				((SlotMixin) (Object) creativeSlots.get(i)).setXPosition(Integer.MIN_VALUE);
 			}
 		}
 	}
 
 	@Inject(at = @At(value = "TAIL"), method = "tick")
 	protected void tick(CallbackInfo info) {
-		float relX = this.mouseX - this.x;
-		float relY = this.mouseY - this.y;
+		if (selectedTab != ItemGroup.INVENTORY.getIndex()) return;
+		List<TrinketSlots.Slot> trinketSlots = TrinketSlots.getAllSlots();
+		float relX = mouseX - this.x;
+		float relY = mouseY - this.y;
 		if (TrinketsClient.slotGroup == null || !inBounds(TrinketsClient.slotGroup, relX, relY, true)) {
 			if (TrinketsClient.slotGroup != null) {
-				for (TrinketSlot ts: invSlots) {
-					if (ts.group.equals(TrinketsClient.slotGroup.getName()) && !ts.keepVisible) {
-						((SlotMixin) (Object) ts).setXPosition(Integer.MIN_VALUE);
+				for (int i = 0; i < trinketSlots.size(); i++) {
+					if (trinketSlots.get(i).getSlotGroup().getName().equals(TrinketsClient.slotGroup.getName())
+							&& (TrinketsClient.slotGroup.onReal || TrinketsClient.slotGroup.slots.get(0) != trinketSlots.get(i))) {
+						((SlotMixin) (Object) creativeSlots.get(i)).setXPosition(Integer.MIN_VALUE);
 					}
 				}
 			}
 			TrinketsClient.slotGroup = null;
-			for (SlotGroup group: TrinketSlots.slotGroups) {
+			for (SlotGroup group : TrinketSlots.slotGroups) {
 				if (inBounds(group, relX, relY, false) && group.slots.size() > 0) {
 					TrinketsClient.displayEquipped = 0;
 					TrinketsClient.slotGroup = group;
-					List<TrinketSlot> tSlots = new ArrayList<TrinketSlot>();
-					for (TrinketSlot ts: invSlots) {
-						if(ts.group.equals(group.getName())) tSlots.add(ts);
+					List<Slot> tSlots = new ArrayList<Slot>();
+					for (int i = 0; i < trinketSlots.size(); i++) {
+						if (trinketSlots.get(i).getSlotGroup() == group)
+							tSlots.add(creativeSlots.get(i));
 					}
+					if (tSlots.size() == 0) return;
 					int groupX = getGroupX(group);
 					int groupY = getGroupY(group);
 					int count = group.slots.size();
@@ -107,7 +105,6 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 					}
 					int l = count / 2;
 					int r = count - l - 1;
-					if(tSlots.size() == 0) break;
 					for (int i = 0; i < l; i++) {
 						((SlotMixin) (Object) tSlots.get(i + offset)).setXPosition(groupX - (i + 1) * 18 + 1);
 						((SlotMixin) (Object) tSlots.get(i + offset)).setYPosition(groupY + 1);
@@ -120,7 +117,7 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 					if (group.vanillaSlot != -1) {
 						TrinketsClient.activeSlots.add(this.getScreenHandler().getSlot(group.vanillaSlot));
 					}
-					for (TrinketSlot ts: tSlots) {
+					for (Slot ts : tSlots) {
 						TrinketsClient.activeSlots.add(ts);
 					}
 					break;
@@ -132,10 +129,13 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 			if (TrinketsClient.slotGroup == null) {
 				SlotGroup group = TrinketsClient.lastEquipped;
 				if (group != null) {
-					List<TrinketSlot> tSlots = new ArrayList<TrinketSlot>();
-					for (TrinketSlot ts: invSlots) {
-						if (ts.group == group.getName()) tSlots.add(ts);
+					List<Slot> tSlots = new ArrayList<Slot>();
+					for (int i = 0; i < trinketSlots.size(); i++) {
+						if (trinketSlots.get(i).getSlotGroup().getName().equals(group.getName())) {
+							tSlots.add(creativeSlots.get(i));
+						}
 					}
+					if (tSlots.size() == 0) return;
 					int groupX = getGroupX(group);
 					int groupY = getGroupY(group);
 					int count = group.slots.size();
@@ -161,73 +161,82 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 					if (group.vanillaSlot != -1) {
 						TrinketsClient.activeSlots.add(this.getScreenHandler().getSlot(group.vanillaSlot));
 					}
-					for (TrinketSlot ts: tSlots) {
+					for (Slot ts : tSlots) {
 						TrinketsClient.activeSlots.add(ts);
 					}
 				}
 			}
 		}
-		for (TrinketSlot ts: invSlots) {
-			if (((TrinketsClient.lastEquipped == null || TrinketsClient.displayEquipped <= 0 || ts.group != TrinketsClient.lastEquipped.getName())
-					&& (TrinketsClient.slotGroup == null || ts.group != TrinketsClient.slotGroup.getName())) && !ts.keepVisible) {
-				((SlotMixin) (Object) ts).setXPosition(Integer.MIN_VALUE);
-			}
-		}
-		for (TrinketSlot ts : invSlots) {
-			int groupX = getGroupX(TrinketSlots.getSlotFromName(ts.group, ts.slot).getSlotGroup());
-			if (ts.keepVisible && groupX < 0) {
-				if (getRecipeBookWidget().isOpen()) {
-					((SlotMixin) (Object) ts).setXPosition(Integer.MIN_VALUE);
-				} else {
-					((SlotMixin) (Object) ts).setXPosition(groupX + 1);
-				}
+		for (int i = 0; i < trinketSlots.size(); i++) {
+			if (((TrinketsClient.lastEquipped == null || TrinketsClient.displayEquipped <= 0 || !trinketSlots.get(i).getSlotGroup().getName().equals(TrinketsClient.lastEquipped.getName()))
+					&& (TrinketsClient.slotGroup == null || !trinketSlots.get(i).getSlotGroup().equals(TrinketsClient.slotGroup)))
+					&& (trinketSlots.get(i).getSlotGroup().onReal || trinketSlots.get(i).getSlotGroup().slots.get(0) != trinketSlots.get(i))) {
+				((SlotMixin) (Object) creativeSlots.get(i)).setXPosition(Integer.MIN_VALUE);
 			}
 		}
 	}
 
-	@Inject(at = @At("TAIL"), method = "drawBackground")
+	@Inject(at = @At(value = "TAIL"), method = "drawBackground")
 	protected void drawBackground(MatrixStack matrices, float f, int x, int y, CallbackInfo info) {
+		if (selectedTab != ItemGroup.INVENTORY.getIndex()) return;
+		GlStateManager.disableLighting();
 		SlotGroup lastGroup = TrinketSlots.slotGroups.get(TrinketSlots.slotGroups.size() - 1);
 		int lastX = getGroupX(lastGroup);
 		int lastY = getGroupY(lastGroup);
-		if (!getRecipeBookWidget().isOpen() && lastX < 0) {
+		if (lastX < 0) {
 			TrinketInventoryRenderer.renderExcessSlotGroups(matrices, this, this.client.getTextureManager(), this.x, this.y, lastX, lastY);
-		}
-		for (SlotGroup group: TrinketSlots.slotGroups) {
-			if (!group.onReal && group.slots.size() > 0) {
-				if (getRecipeBookWidget().isOpen() && getGroupX(group) < 0) continue;
-				this.client.getTextureManager().bindTexture(TrinketInventoryRenderer.MORE_SLOTS_TEX);
-				this.drawTexture(matrices, this.x + getGroupX(group), this.y + getGroupY(group), 4, 4, 18, 18);
-			}
 		}
 	}
 
 	@Inject(at = @At(value = "TAIL"), method = "drawForeground")
 	protected void drawForeground(MatrixStack matrices, int x, int y, CallbackInfo info) {
+		if (selectedTab != ItemGroup.INVENTORY.getIndex()) return;
 		if (TrinketsClient.slotGroup != null) {
-			TrinketInventoryRenderer.renderGroupFront(matrices, this, this.client.getTextureManager(), this.playerInventory, 0, 0, TrinketsClient.slotGroup, getGroupX(TrinketsClient.slotGroup), getGroupY(TrinketsClient.slotGroup));
+			TrinketInventoryRenderer.renderGroupFront(matrices, this, this.client.getTextureManager(), this.playerInventory,
+					0, 0, TrinketsClient.slotGroup, getGroupX(TrinketsClient.slotGroup),
+					getGroupY(TrinketsClient.slotGroup));
 		} else if (TrinketsClient.displayEquipped > 0 && TrinketsClient.lastEquipped != null) {
-			TrinketInventoryRenderer.renderGroupFront(matrices, this, this.client.getTextureManager(), this.playerInventory, 0, 0, TrinketsClient.lastEquipped, getGroupX(TrinketsClient.lastEquipped), getGroupY(TrinketsClient.lastEquipped));
+			TrinketInventoryRenderer.renderGroupFront(matrices, this, this.client.getTextureManager(), this.playerInventory,
+					0, 0, TrinketsClient.lastEquipped, getGroupX(TrinketsClient.lastEquipped),
+					getGroupY(TrinketsClient.lastEquipped));
 		} else {
 			return;
 		}
 		super.drawForeground(matrices, x, y);
 		RenderSystem.disableLighting();
+		for (SlotGroup group : TrinketSlots.slotGroups) {
+			if (!group.onReal && group.slots.size() > 0) {
+				this.client.getTextureManager().bindTexture(TrinketInventoryRenderer.MORE_SLOTS_TEX);
+				this.drawTexture(matrices, getGroupX(group), getGroupY(group), 4, 4, 18, 18);
+			}
+		}
 	}
 
-	@Inject(at = @At("HEAD"), method = "isClickOutsideBounds", cancellable = true)
-	protected void isClickOutsideBounds(double x, double y, int i, int j, int k, CallbackInfoReturnable<Boolean> info) {
-		if (TrinketsClient.slotGroup != null && inBounds(TrinketsClient.slotGroup, (float) x - this.x, (float) y - this.y, true)) info.setReturnValue(false);
+	@Inject(at = @At(value = "TAIL"), method = "render")
+	protected void render(MatrixStack matrices, int x, int y, float f, CallbackInfo info) {
+		if (selectedTab != ItemGroup.INVENTORY.getIndex()) return;
+		mouseX = x;
+		mouseY = y;
+	}
+
+	@Inject(at = @At(value = "TAIL"), method = "setSelectedTab")
+	private void setSelectedTab(ItemGroup itemGroup, CallbackInfo info) {
+		if (itemGroup == ItemGroup.INVENTORY) {
+			ScreenHandler screenHandler = this.getScreenHandler();
+			creativeSlots.clear();
+			for (int i = 46; i < screenHandler.slots.size(); i++) {
+				if(screenHandler.getSlot(i).inventory instanceof TrinketInventory) creativeSlots.add(screenHandler.getSlot(i));
+			}
+			init(info);
+		}
 	}
 
 	public boolean inBounds(SlotGroup group, float x, float y, boolean focused) {
-		if (getRecipeBookWidget().isOpen() && getGroupX(group) < 0) return false;
 		int groupX = getGroupX(group);
 		int groupY = getGroupY(group);
 		if (focused) {
 			int count = group.slots.size();
-			if (group.onReal)
-				count++;
+			if (group.onReal) count++;
 			int l = count / 2;
 			int r = count - l - 1;
 			return x > groupX - l * 18 - 4 && y > groupY - 4 && x < groupX + r * 18 + 22 && y < groupY + 22;
@@ -237,18 +246,20 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 	}
 
 	public int getGroupX(SlotGroup group) {
-		if (group.vanillaSlot == 5) return 7;
-		if (group.vanillaSlot == 6) return 7;
-		if (group.vanillaSlot == 7) return 7;
-		if (group.vanillaSlot == 8) return 7;
-		if (group.vanillaSlot == 45) return 76;
-		if (group.getName().equals(SlotGroups.HAND)) return 76;
+		if (group.vanillaSlot == 5) return 53;
+		if (group.vanillaSlot == 6) return 53;
+		if (group.vanillaSlot == 7) return 107;
+		if (group.vanillaSlot == 8) return 107;
+		if (group.vanillaSlot == 45) return 34;
+		if (group.getName().equals("hand")) return 15;
 		int j = 0;
 		if (TrinketSlots.slotGroups.get(5).slots.size() == 0) j = -1;
 		for (int i = 6; i < TrinketSlots.slotGroups.size(); i++) {
 			if (TrinketSlots.slotGroups.get(i) == group) {
 				j += i;
-				if (j < 8) return 76;
+				if (j == 5) return 15;
+				if (j == 6) return 126;
+				if (j == 7) return 145;
 				return -15 - ((j - 8) / 4) * 18;
 			} else if (TrinketSlots.slotGroups.get(i).slots.size() == 0) j--;
 		}
@@ -256,20 +267,18 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 	}
 
 	public int getGroupY(SlotGroup group) {
-		if (group.vanillaSlot == 5) return 7;
-		if (group.vanillaSlot == 6) return 25;
-		if (group.vanillaSlot == 7) return 43;
-		if (group.vanillaSlot == 8) return 61;
-		if (group.vanillaSlot == 45) return 61;
-		if (group.getName().equals(SlotGroups.HAND)) return 43;
+		if (group.vanillaSlot == 5) return 5;
+		if (group.vanillaSlot == 6) return 32;
+		if (group.vanillaSlot == 7) return 5;
+		if (group.vanillaSlot == 8) return 32;
+		if (group.vanillaSlot == 45) return 19;
+		if (group.getName().equals("hand")) return 19;
 		int j = 0;
 		if (TrinketSlots.slotGroups.get(5).slots.size() == 0) j = -1;
 		for (int i = 6; i < TrinketSlots.slotGroups.size(); i++) {
 			if (TrinketSlots.slotGroups.get(i) == group) {
 				j += i;
-				if (j == 5) return 43;
-				if (j == 6) return 25;
-				if (j == 7) return 7;
+				if (j < 8) return 19;
 				return 7 + ((j - 8) % 4) * 18;
 			} else if (TrinketSlots.slotGroups.get(i).slots.size() == 0) j--;
 		}
