+++ src/main/java/dev/emi/trinkets/mixin/client/HandledScreenMixin.java
@@ -1,4 +1,4 @@
-package dev.emi.trinkets.mixin;
+package dev.emi.trinkets.mixin.client;
 
 import java.util.List;
